requirements.txt
aio-pika==9.5.8
aiormq==6.9.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.3.0
minio==7.2.19
multidict==6.7.0
packaging==25.0
pamqp==3.3.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
